A few changes - Implement setting to auto-stop resource usage manager when no one is online to save bandwidth - Refactor: rename ResourceUsageManager#initializeListener() to startListener() - Refactor: move event listener to its own package and rename the class
